// Frontend/Frontend-main/p2-frontend/src/app/models/User.ts
// Created file so folder is tracked in git
export class User {
  userId: number|undefined;
  username!: string|undefined;
  password: string|undefined;
  email: string|undefined;
  firstName: string|undefined;
  lastName: string|undefined;
  aboutMe: string| undefined;
  bday: Date| undefined;
  proPicUrl: string|undefined;
}